+++ OnlineAdmission.APP/ViewModels/Student/StudentVM.cs
@@ -92,7 +92,7 @@ namespace OnlineAdmission.APP.ViewModels.Student
         public string Religion { get; set; }
 
         [Display(Name = "H.S.C Roll"), Required(ErrorMessage = "HSC Roll Required")]
-        public int HSCRoll { get; set; }
+        public long HSCRoll { get; set; }
 
         [Display(Name = "S.S.C Roll"), Required(ErrorMessage = "SSC Roll Required")]
         public int SSCRoll { get; set; }
