+++ OnlineAdmission.DAL/IRepository/IStudentRepository.cs
@@ -9,7 +9,7 @@ namespace OnlineAdmission.DAL.IRepository
 {
     public interface IStudentRepository : IRepository<Student>
     {
-        Task<Student> GetStudentByHSCRollAsync(int hscRoll);
+        Task<Student> GetStudentByHSCRollAsync(long hscRoll);
         Task<Student> GetStudentBySSCRollAsync(int sscRoll, string boardName);
         Task<int> GetCountAsync(int subId);
         Task<Student> GetByAdmissionRollAsync(int NURoll);
